tests/wpt/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-presentational-hint.html

<!doctype html>
<title>CSS Color Adjust Test: meta color-scheme - presentational hint</title>
<link rel="help" href="https://drafts.csswg.org/css-color-adjust-1/#color-scheme-meta">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/compute-root-color-scheme.js"></script>
<style>
  * { color-scheme: light }
</style>
<meta name="color-scheme" content="dark">
<script>
  // Author level CSS color-scheme property overrides meta color-scheme value
  // which is at the presentational hint cascade order.
  assert_root_color_scheme("light", "Author origin style rule overrides meta color-scheme.");
</script>
