third_party/blink/web_tests/external/wpt/css/css-color-adjust/meta/meta-color-scheme-presentational-hint-expected.txt

This is a testharness.js-based test.
FAIL Author origin style rule overrides meta color-scheme. assert_equals: expected (string) "light" but got (undefined) undefined
Harness: the test ran to completion.
